fix size handling

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -9,7 +9,7 @@
 
 int get_bit(unsigned long int n, unsigned int index)
 {
-	unsigned int i, size = sizeof(unsigned long int) * 8;
+	unsigned long int i, size = sizeof(unsigned long int) * 8;
 	int msbp, result = -1;/*msp-most significant bit position*/
 
 	for (msbp = size - 1; msbp >= 0; msbp--)
